feat(blog): refactor ID handling to use uint types for improved data integrity

+++ internal/app/service/blog/blog.go
@@ -8,7 +8,7 @@ import (
 )
 
 func isPermission(b blog.Blog, reqUser model.ReqUser) error {
-	if b.UserId != uint64(reqUser.Id) && reqUser.Role < entity.RoleAdmin {
+	if b.UserId != reqUser.Id && reqUser.Role < entity.RoleAdmin {
 		return &errors.ErrUnauthorized
 	}
 	return nil
